backend/requirements.txt
# AgentKit SaaS — Backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0.1
python-multipart>=0.0.6

# Async DB drivers
aiosqlite>=0.19.0
asyncpg>=0.29.0
greenlet>=3.0.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0

# Migrations
alembic>=1.13.0

# Email validation
pydantic[email]>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
